Remove OSGi dependencies form http transport
