+++ jaggr-service/.settings/.gitignore
@@ -1,2 +1,3 @@
 /org.eclipse.core.resources.prefs
 /org.eclipse.m2e.core.prefs
+/de.loskutov.FileSync.prefs
