+++ jaggr-service/src/test/java/com/ibm/jaggr/service/impl/transport/AbstractDojoHttpTransportTest.java
@@ -40,7 +40,7 @@ import com.ibm.jaggr.core.modulebuilder.IModuleBuilderExtensionPoint;
 import com.ibm.jaggr.service.impl.config.ConfigImpl;
 import com.ibm.jaggr.service.test.TestUtils;
 
-public class DojoHttpTransportTest {
+public class AbstractDojoHttpTransportTest {
 
 	@BeforeClass
 	public static void setUpBeforeClass() throws Exception {
@@ -75,10 +75,7 @@ public class DojoHttpTransportTest {
 		String config = "{packages:[{name:\"dojo\", location:\"namedbundleresource://com.ibm.jaggr.sample.dojo/WebContent/dojo\"}]}";
 		URI tmpDir = new File(System.getProperty("java.io.tmpdir")).toURI();
 		IConfig cfg = new ConfigImpl(mockAggregator, tmpDir, config);
-		final URI comboUri = new URI(DojoHttpTransport.comboUriStr);
-		DojoHttpTransport transport = new DojoHttpTransport() {
-			@Override protected String getResourcePathId() {return "combo";}
-			@Override protected URI getComboUri() { return comboUri; }
+		AbstractDojoHttpTransport transport = new TestUtils.TestDojoHttpTransport() {
 			@Override public List<String[]> getClientConfigAliases() {
 				return super.getClientConfigAliases();
 			}
@@ -90,38 +87,38 @@ public class DojoHttpTransportTest {
 		System.out.println(Context.toString(rawConfig));
 		// verify config has the changes we expect
 		modifiedConfig = new ConfigImpl(
-				mockAggregator, 
+				mockAggregator,
 				new File(System.getProperty("java.io.tmpdir")).toURI(),
 				rawConfig);
 		Assert.assertEquals(3, modifiedConfig.getPaths().size());
 		Assert.assertEquals("namedbundleresource://com.ibm.jaggr.sample.dojo/WebContent/dojo/text",
-				modifiedConfig.getPaths().get(DojoHttpTransport.dojoTextPluginAliasFullPath).getPrimary().toString());
-		Assert.assertNull(modifiedConfig.getPaths().get(DojoHttpTransport.dojoTextPluginAliasFullPath).getOverride());
-		Assert.assertEquals(DojoHttpTransport.textPluginProxyUriStr,
-				modifiedConfig.getPaths().get(DojoHttpTransport.dojoTextPluginFullPath).getPrimary().toString());
-		Assert.assertEquals(comboUri, modifiedConfig.getPaths().get("combo").getPrimary());
+				modifiedConfig.getPaths().get(AbstractDojoHttpTransport.dojoTextPluginAliasFullPath).getPrimary().toString());
+		Assert.assertNull(modifiedConfig.getPaths().get(AbstractDojoHttpTransport.dojoTextPluginAliasFullPath).getOverride());
+		Assert.assertEquals(transport.getComboUri().resolve(AbstractDojoHttpTransport.textPluginPath).toString(),
+				modifiedConfig.getPaths().get(AbstractDojoHttpTransport.dojoTextPluginFullPath).getPrimary().toString());
+		Assert.assertEquals(transport.getComboUri(), modifiedConfig.getPaths().get("combo").getPrimary());
 		Assert.assertEquals(1, modifiedConfig.getAliases().size());
-		Assert.assertEquals(DojoHttpTransport.dojoTextPluginAlias, modifiedConfig.getAliases().get(0).getPattern());
-		Assert.assertEquals(DojoHttpTransport.dojoTextPluginAliasFullPath, modifiedConfig.getAliases().get(0).getReplacement());
+		Assert.assertEquals(AbstractDojoHttpTransport.dojoTextPluginAlias, modifiedConfig.getAliases().get(0).getPattern());
+		Assert.assertEquals(AbstractDojoHttpTransport.dojoTextPluginAliasFullPath, modifiedConfig.getAliases().get(0).getReplacement());
 		Assert.assertEquals(2, transport.getClientConfigAliases().size());
 		String[] alias = transport.getClientConfigAliases().get(0);
-		Assert.assertEquals(DojoHttpTransport.dojoTextPluginAlias, alias[0]);
-		Assert.assertEquals(DojoHttpTransport.dojoTextPluginAliasFullPath, alias[1]);
+		Assert.assertEquals(AbstractDojoHttpTransport.dojoTextPluginAlias, alias[0]);
+		Assert.assertEquals(AbstractDojoHttpTransport.dojoTextPluginAliasFullPath, alias[1]);
 		alias = transport.getClientConfigAliases().get(1);
-		Assert.assertEquals(DojoHttpTransport.aggregatorTextPluginAlias, alias[0]);
+		Assert.assertEquals(AbstractDojoHttpTransport.aggregatorTextPluginAlias, alias[0]);
 		Assert.assertEquals("combo/text", alias[1]);
-		
+
 		// make sure overrides are handled properly
 		config = "{packages:[{name:\"dojo\", location:[\"namedbundleresource://com.ibm.jaggr.sample.dojo/WebContent/dojo\"]}]}";
 		cfg = new ConfigImpl(mockAggregator, tmpDir, config);
 		rawConfig = cfg.getRawConfig();
 		transport.modifyConfig(mockAggregator, rawConfig);
 		modifiedConfig = new ConfigImpl(
-				mockAggregator, 
+				mockAggregator,
 				new File(System.getProperty("java.io.tmpdir")).toURI(),
 				rawConfig);
 		System.out.println(Context.toString(rawConfig));
-		IConfig.Location loc = modifiedConfig.getPaths().get(DojoHttpTransport.dojoTextPluginAliasFullPath);
+		IConfig.Location loc = modifiedConfig.getPaths().get(AbstractDojoHttpTransport.dojoTextPluginAliasFullPath);
 		Assert.assertEquals("namedbundleresource://com.ibm.jaggr.sample.dojo/WebContent/dojo/text", loc.getPrimary().toString());
 		Assert.assertNull(loc.getOverride());
 
@@ -130,29 +127,29 @@ public class DojoHttpTransportTest {
 		rawConfig = cfg.getRawConfig();
 		transport.modifyConfig(mockAggregator, rawConfig);
 		modifiedConfig = new ConfigImpl(
-				mockAggregator, 
+				mockAggregator,
 				new File(System.getProperty("java.io.tmpdir")).toURI(),
 				rawConfig);
 		System.out.println(Context.toString(rawConfig));
-		loc = modifiedConfig.getPaths().get(DojoHttpTransport.dojoTextPluginAliasFullPath);
+		loc = modifiedConfig.getPaths().get(AbstractDojoHttpTransport.dojoTextPluginAliasFullPath);
 		Assert.assertEquals("namedbundleresource://com.ibm.jaggr.sample.dojo/WebContent/dojo/text", loc.getPrimary().toString());
 		Assert.assertEquals("file:/c:/customizations/dojo/text", loc.getOverride().toString());
-		
+
 		// No dojo package in config
 		transport.getClientConfigAliases().clear();
 		cfg = new ConfigImpl(mockAggregator, tmpDir, "{}");
 		rawConfig = cfg.getRawConfig();
 		transport.modifyConfig(mockAggregator, rawConfig);
 		modifiedConfig = new ConfigImpl(
-				mockAggregator, 
+				mockAggregator,
 				new File(System.getProperty("java.io.tmpdir")).toURI(),
 				rawConfig);
 		System.out.println(Context.toString(rawConfig));
 		Assert.assertEquals(1, modifiedConfig.getPaths().size());
 		Assert.assertEquals(0, modifiedConfig.getAliases().size());
-		Assert.assertEquals(comboUri, modifiedConfig.getPaths().get("combo").getPrimary());
+		Assert.assertEquals(transport.getComboUri(), modifiedConfig.getPaths().get("combo").getPrimary());
 		Assert.assertEquals(0, transport.getClientConfigAliases().size());
-		
+
 		// More that one text module builder.  Make sure last one is used
 		extensions.add(mockExtension);
 		mockExtension = EasyMock.createNiceMock(IAggregatorExtension.class);
@@ -163,25 +160,24 @@ public class DojoHttpTransportTest {
 		rawConfig = cfg.getRawConfig();
 		transport.modifyConfig(mockAggregator, rawConfig);
 		modifiedConfig = new ConfigImpl(
-				mockAggregator, 
+				mockAggregator,
 				new File(System.getProperty("java.io.tmpdir")).toURI(),
 				rawConfig);
 		System.out.println(Context.toString(rawConfig));
 		Assert.assertEquals(3, modifiedConfig.getPaths().size());
 		Assert.assertEquals(1, modifiedConfig.getAliases().size());
-		Assert.assertEquals(comboUri, modifiedConfig.getPaths().get("combo").getPrimary());
+		Assert.assertEquals(transport.getComboUri(), modifiedConfig.getPaths().get("combo").getPrimary());
 		Assert.assertEquals("namedbundleresource://com.ibm.jaggr.sample.dojo/WebContent/dojo/text",
-				modifiedConfig.getPaths().get(DojoHttpTransport.dojoTextPluginAliasFullPath).getPrimary().toString());
-		Assert.assertEquals(DojoHttpTransport.textPluginProxyUriStr,
-				modifiedConfig.getPaths().get(DojoHttpTransport.dojoTextPluginFullPath).getPrimary().toString());
+				modifiedConfig.getPaths().get(AbstractDojoHttpTransport.dojoTextPluginAliasFullPath).getPrimary().toString());
+		Assert.assertEquals(transport.getComboUri().resolve(AbstractDojoHttpTransport.textPluginPath).toString(),
+				modifiedConfig.getPaths().get(AbstractDojoHttpTransport.dojoTextPluginFullPath).getPrimary().toString());
 		Assert.assertEquals(2, transport.getClientConfigAliases().size());
 		alias = transport.getClientConfigAliases().get(0);
-		Assert.assertEquals(DojoHttpTransport.dojoTextPluginAlias, alias[0]);
-		Assert.assertEquals(DojoHttpTransport.dojoTextPluginAliasFullPath, alias[1]);
+		Assert.assertEquals(AbstractDojoHttpTransport.dojoTextPluginAlias, alias[0]);
+		Assert.assertEquals(AbstractDojoHttpTransport.dojoTextPluginAliasFullPath, alias[1]);
 		alias = transport.getClientConfigAliases().get(1);
-		Assert.assertEquals(DojoHttpTransport.aggregatorTextPluginAlias, alias[0]);
+		Assert.assertEquals(AbstractDojoHttpTransport.aggregatorTextPluginAlias, alias[0]);
 		Assert.assertEquals("combo/text", alias[1]);
 		Context.exit();
 	}
-
 }
